% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{clbic}
\alias{clbic}
\alias{clbic.mvord}
\title{CLBIC of Multivariate Ordinal Regression Models.}
\usage{
clbic(object)

\method{clbic}{mvord}(object)
}
\arguments{
\item{object}{object of class \code{"mvord"}}
}
\description{
\code{clbic} is a generic function which extracts the composite likelihood BIC from objects of class \cr
\code{"mvord"}.
}
