% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{coef.multord}
\alias{coef.multord}
\title{Coefficients of Multivariate Ordinal Regression Models.}
\usage{
\method{coef}{multord}(object, ...)
}
\arguments{
\item{object}{object of class \code{"multord"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{coef} is a generic function which extracts
 regression coefficients from objects of class \code{"multord"}.
}
