% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{coef.mvord}
\alias{coef.mvord}
\title{Coefficients of Multivariate Ordinal Regression Models.}
\usage{
\method{coef}{mvord}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{coef} is a generic function which extracts
 regression coefficients from objects of class \code{'mvord'}.
}
