% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{constraints}
\alias{constraints}
\alias{constraints.mvord}
\title{Regression constraints of Multivariate Ordinal Regression Models.}
\usage{
constraints(object)

\method{constraints}{mvord}(object)
}
\arguments{
\item{object}{object of class \code{"mvord"}}
}
\description{
\code{constraints} is a generic function which extracts the regression constraints from objects of class \cr
\code{"mvord"}.
}
