% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_struct.R
\name{corr_jac_num_fct}
\alias{corr_jac_num_fct}
\title{corr_jac.cor_rel_var <- function(eobj, tpar){
  ndim <- attr(eobj, "ndim")
  npar.cor <- attr(eobj, "npar.cor")
  npar.sd <- attr(eobj, "npar.sd")
  l <- list(sapply(1:npar.cor, function(i) grad(function(x) corr_jac_num_fct(ndim, x, i),
                                                x=tpar[seq_len(npar.cor)])))
  l[1 + seq_len(npar.sd)] <- exp(tpar[npar.cor + seq_len(npar.sd)])
  l
}}
\usage{
corr_jac_num_fct(ndim, nu, i)
}
\description{
corr_jac.cor_rel_var <- function(eobj, tpar){
  ndim <- attr(eobj, "ndim")
  npar.cor <- attr(eobj, "npar.cor")
  npar.sd <- attr(eobj, "npar.sd")
  l <- list(sapply(1:npar.cor, function(i) grad(function(x) corr_jac_num_fct(ndim, x, i),
                                                x=tpar[seq_len(npar.cor)])))
  l[1 + seq_len(npar.sd)] <- exp(tpar[npar.cor + seq_len(npar.sd)])
  l
}
}
