% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\docType{data}
\name{data_cr}
\alias{data_cr}
\title{Simulated credit ratings}
\format{A data frame with 690 rows and 11 columns}
\usage{
data(data_cr)
}
\description{
A data set containing simulated credit ratings and simulated perfomance measures from four raters.
}
\details{
\itemize{
  \item \code{rater1} credit ratings assigned by rater 1
  \item \code{rater2} credit ratings assigned by rater 2
  \item \code{rater3} credit ratings assigned by rater 3
  \item \code{rater4} credit ratings assigned by rater 4
  \item \code{firm_id} firm index
  \item \code{rater_id} rater index
covered from the free operating cash-flow of a company
  \item \code{LR} liquidity ratio, relating the cash held by a company to the current liabilities
  \item \code{LEV} leverage ratio relating debt to earnings before interest and taxes
  \item \code{PR} profitability ratio of retained earnoings to assets
  \item \code{RSIZE} log of relative size of the company in the market
  \item \code{BETA} a measure of systematic risk
}
}
