% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\docType{data}
\name{data_cr_multord}
\alias{data_cr_multord}
\title{Simulated credit ratings}
\format{A data frame with 4566 rows and 11 variables}
\usage{
data(data_cr_multord)
}
\description{
A data set containing simulated credit ratings and simulated perfomance measures from four raters.
}
\details{
\itemize{
  \item \code{rating} credit ratings
  \item \code{firm_id} firm index
  \item \code{rater_id} rater index
  \item \code{ICR} interest coverage ratio, which measures how well the interest expenses can be
covered from the free operating cash-flow of a company
  \item \code{LR} liquidity ratio, relating the cash held by a company to the current liabilities
  \item \code{LEV1} leverage ratio relating debt to earnings before interest and taxes
  \item \code{LEV2} leverage ratio measuring the percentage of debt in the long-term capital of a firm
  \item \code{PR} profitability ratio measuring return on capital
  \item \code{lRSIZE} log of relative size of the company in the market
  \item \code{lSYSR} log of a measure of systematic risk
  \item \code{BSEC} business sector of a firm (factor with 8 levels)
}
}
