% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\docType{data}
\name{data_mvord_toy}
\alias{data_mvord_toy}
\title{Data set toy example}
\format{A data frame with 100 rows and 6 variables}
\usage{
data("data_mvord_toy", package = "mvord")
}
\description{
A data set containing two simulated ordinal responses with three categories,
two quantitative covariates \code{X1} and \code{X2} and two categorical covariates
\code{f1} and \code{f2}.
}
\details{
\itemize{
  \item \code{Y1} ordinal outcome \code{Y1} (three categories)
  \item \code{Y2} ordinal outcome \code{Y2} (three categories)
  \item \code{X1} covariate \code{X1}
  \item \code{X2} covariate \code{X2}
  \item \code{f1} categorical covariate \code{f1}
  \item \code{f2} categorical covariate \code{f2}
}
}
