% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\docType{data}
\name{data_toy_example}
\alias{data_toy_example}
\title{Data set toy example}
\format{A data frame with 100 rows and 4 variables}
\usage{
data(data_toy_example)
}
\description{
A data set containing two simulated ordinal observations with three categories and two covariates \code{X1} and \code{X2}.
}
\details{
\itemize{
  \item \code{Y1} ordinal observation one \code{Y1} (three categories)
  \item \code{Y2} ordinal observation one \code{Y2} (three categories)
  \item \code{X1} covariate \code{X1}
  \item \code{X2} covariate \code{X2}
}
}
