% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error.structures}
\alias{error.structures}
\alias{corGeneral}
\alias{corGeneral.formula}
\alias{corEqui}
\alias{corEqui.formula}
\alias{covGeneral}
\alias{covGeneral.formula}
\alias{corAR1}
\alias{corAR1.formula}
\title{Error Structures in mvord}
\usage{
corGeneral(formula)

\method{corGeneral}{formula}(formula)

corEqui(formula)

\method{corEqui}{formula}(formula)

covGeneral(formula)

\method{covGeneral}{formula}(formula)

corAR1(formula)

\method{corAR1}{formula}(formula)
}
\arguments{
\item{formula}{\code{\link{formula}} object}
}
\description{
Different \code{error.structures} are available in \pkg{mvord}:
\itemize{
\item general correlation structure (default) \code{corGeneral(~1)},
\item general covariance structure \code{covGeneral(~1)},
\item factor dependent correlation structure \code{covGeneral(~f)},
\item factor dependent covariance structure \code{covGeneral(~f)},
\item covariate dependent equicorrelation structure \code{corEqui(~S)},
\item AR(1) correlation structure \code{corAR1(~1)}, or
\item covariate dependent AR(1) correlation structure \code{corAR1(~S)}.
}
See \code{\link{error.structures}} or vignette.
}
