% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_struct.R
\name{error_struct}
\alias{error_struct}
\alias{cov_general}
\alias{cor_general}
\alias{cor_ar1}
\alias{cor_equi}
\alias{cor_ident}
\title{Error Structures in mvord}
\usage{
cov_general(formula = ~1)

cor_general(formula = ~1)

cor_ar1(formula = ~1)

cor_equi(formula = ~1)

cor_ident(formula = ~1)
}
\arguments{
\item{formula}{\code{\link{formula}} object}
}
\description{
Different \code{error.structures} are available in \pkg{mvord}:
\itemize{
\item general correlation structure (default) \code{cor_general(~ 1)},
\item general covariance structure \code{cov_general(~ 1)},
\item factor dependent correlation structure \code{cor_general(~ f)},
\item factor dependent covariance structure \code{cov_general(~ f)},
\item equicorrelation structure \code{cor_equi(~ 1)},
\item covariate dependent equicorrelation structure \code{cor_equi(~ S)},
\item AR(1) correlation structure \code{cor_ar1(~ 1)}, or
\item covariate dependent AR(1) correlation structure \code{cor_ar1(~ S)}.
}
For more details see vignette.
}
