% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\docType{data}
\name{essay_data}
\alias{essay_data}
\title{Essay data}
\format{
A data frame with 198 rows and 6 variables
}
\usage{
data("essay_data", package = "mvord")
}
\description{
The multirater agreement data set is taken from Chapter 5 in "Ordinal Data Modeling",
from Johnson, Valen E and Albert, J. The data consists of grades assigned to
198 essays by 5 experts, each of whom rated all essays on a 10-point
scale. A score of 10 indicates an excellent essay. In addition, the
average word length is also available as an essay characteristic.
}
\details{
\itemize{
  \item \code{Judge1} ordinal outcome:  grades assigned by expert 1
  \item \code{Judge2} ordinal outcome:  grades assigned by expert 2
  \item \code{Judge3} ordinal outcome:  grades assigned by expert 3
  \item \code{Judge4} ordinal outcome:  grades assigned by expert 4
  \item \code{Judge5} ordinal outcome:  grades assigned by expert 5
  \item \code{wl} covariate: word length
}
}
