% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{fitted.mvord}
\alias{fitted.mvord}
\title{Fitted Probabilities of Multivariate Ordinal Regression Models.}
\usage{
\method{fitted}{mvord}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A generic function which extracts fitted probabilities for the observed categories from objects of class
\code{'mvord'}.
}
