% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get.prob}
\alias{get.prob}
\title{Extracts fitted Probabilities for Multivariate Ordinal Regression Models.}
\usage{
get.prob(object, response.cat, subjectID = NULL)
}
\arguments{
\item{object}{of class \code{mvord}}

\item{response.cat}{vector or matrix with response categories (for each subject one row with q multiple measurements).}

\item{subjectID}{(optional) vector specifying for which subjectIDs the predictions\cr or fitted values should be computed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts fitted probabilities for given response categories from a fitted model of class \code{"mvord"}.
}
\details{
The current implemntation supports only in-sample predictions. 
The rownames of the output correspond to the subjectIDs.
}
\seealso{
\code{\link{predict.mvord}}, \code{\link{marginal.predict}}
}
