% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{logPL}
\alias{logPL}
\alias{logPL.mvord}
\title{logPL of Multivariate Ordinal Regression Models.}
\usage{
logPL(object)

\method{logPL}{mvord}(object)
}
\arguments{
\item{object}{object of class \code{"mvord"}}
}
\description{
\code{logPL} is a generic function which extracts the log pairwise likelihood from objects of class \cr
\code{"mvord"}.
}
