% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{model.matrix.mvord}
\alias{model.matrix.mvord}
\title{model.matrix of Multivariate Ordinal Regression Models.}
\usage{
\method{model.matrix}{mvord}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{model.matrix} is a generic function which extracts the model matrix
 from objects of class \code{'mvord'}.
}
