% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{mvord-package}
\alias{mvord-package}
\title{Multivariate Ordinal Regression Models in R.}
\description{
The R package mvord implements composite likelihood
estimation in the class of multivariate ordinal regression models with probit and logit link.
A flexible modeling framework for multiple ordinal measurements on
the same subject is set up, which takes into consideration the
dependence among the multiple observations by employing different
error structures.
Heterogeneity in the error structure across the subjects can be
accounted for by the package, which allows for covariate dependent
error structures.
In addition, regression coefficients and threshold parameters are
varying across the multiple response dimensions in the default
implementation. However, constraints can be defined by the user if a
reduction of the parameter space is desired.
}
\details{
see \code{\link{mvord}}, \code{\link{mvord2}}
}
