% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mvord.control}
\alias{mvord.control}
\title{Control functions for mvord()}
\usage{
mvord.control(se = TRUE, start.values = NULL, solver = "newuoa",
  solver.optimx.control = list(maxit = 2e+05, trace = 0, kkt = FALSE))
}
\arguments{
\item{se}{logical, if \code{TRUE} standard errors are computed.}

\item{start.values}{vector of (optional) starting values.}

\item{solver}{character string containing the name of the applicable solver of \code{\link{optimx}} (default is \code{"newuoa"})
or wrapper function for user defined solver.}

\item{solver.optimx.control}{a list of control arguments to be passed to \code{\link{optimx}}. See \code{\link{optimx}}.}
}
\description{
Control arguments are set for \code{mvord()}.
}
\seealso{
\code{\link{mvord}}
}
