% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{names_constraints}
\alias{names_constraints}
\title{Names of regression coefficient constraints in mvord}
\usage{
names_constraints(formula, data, contrasts = NULL)
}
\arguments{
\item{formula}{model formula}

\item{data}{a given data set.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{\link{model.matrix.default}}.}
}
\description{
An extractor function for the names of the regression coefficient constraints based on the model \code{formula} and \code{data}.
}
