% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{nobs.mvord}
\alias{nobs.mvord}
\title{nobs of Multivariate Ordinal Regression Models.}
\usage{
\method{nobs}{mvord}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{nobs} is a generic function which extracts the number of observations from objects of class \cr
\code{'mvord'}.
}
