% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{polycor}
\alias{polycor}
\title{Computes polychoric correlations}
\usage{
polycor(x, y = NULL)
}
\arguments{
\item{x}{either a vector or a matrix of ordinal responses}

\item{y}{an (optional) ordinal vector (only applciable if x is a vector)}
}
\description{
This function computes polychoric correleations among two or more variables.
}
