% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_struct.R
\name{print.error_struct}
\alias{print.error_struct}
\title{Print Method for class error_struc.}
\usage{
\method{print}{error_struct}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{error_struct}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints error structure of class \code{\link{error_struct}}.
}
