% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{print.multord}
\alias{print.multord}
\title{Print Method for Multivariate Ordinal Regression Models.}
\usage{
\method{print}{multord}(x, call = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{"multord"}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints thresholds, regression coefficients
 and parameters of the error structure of class \code{"multord"}.
}
