% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{print.mvord}
\alias{print.mvord}
\title{Print Method for Multivariate Ordinal Regression Models.}
\usage{
\method{print}{mvord}(x, call = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{'mvord'}}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints thresholds, regression coefficients
 and parameters of the error structure of class \code{'mvord'}.
}
