% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pseudo_R_squared}
\alias{pseudo_R_squared}
\title{Pseudo \eqn{R^2} for objects of class 'mvord'}
\usage{
pseudo_R_squared(object, adjusted = FALSE)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}

\item{adjusted}{if \code{TRUE}, then adjusted Mc Fadden's Pseudo \eqn{R^2} is computed.}
}
\description{
This function computes Mc Fadden's Pseudo \eqn{R^2} for objects of class \code{'mvord'}.
}
\seealso{
\code{\link{mvord}}
}
