% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{summary.mvord}
\alias{summary.mvord}
\title{Summary method for Multivariate Ordinal Regression Models.}
\usage{
\method{summary}{mvord}(object, short = TRUE, call = TRUE, ...)
}
\arguments{
\item{object}{object of class \code{'mvord'}}

\item{short}{if \code{TRUE} short summary, otherwise extended summary}

\item{call}{displays function call if \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary of thresholds, regression coefficients
and parameters of the error structure of class \code{'mvord'}.
}
