% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{terms.mvord}
\alias{terms.mvord}
\title{terms of Multivariate Ordinal Regression Models.}
\usage{
\method{terms}{mvord}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'mvord'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{terms} is a generic function which can be used to extract terms from objects of class \cr
\code{'mvord'}.
}
