% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{thresholds}
\alias{thresholds}
\alias{thresholds.multord}
\title{Thresholds of Multivariate Ordinal Regression Models.}
\usage{
thresholds(object, ...)

\method{thresholds}{multord}(object, ...)
}
\arguments{
\item{object}{object of class \code{"multord"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{thresholds} is a generic function which extracts threshold coefficients from objects of class \cr
\code{"multord"}.
}
