% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{vcov.mvord}
\alias{vcov.mvord}
\title{vcov of Multivariate Ordinal Regression Models.}
\usage{
\method{vcov}{mvord}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{vcov} is a generic function which extracts the Godambe information matrix from objects of class \cr
\code{'mvord'}.
}
