% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\name{constraints}
\alias{constraints}
\alias{constraints.mvord}
\title{Constraints on the Regression Coefficients of Multivariate Ordinal Regression Models.}
\usage{
constraints(object)

\method{constraints}{mvord}(object)
}
\arguments{
\item{object}{an object of class \code{'mvord'}.}
}
\description{
\code{constraints} is a generic function which extracts the constraints on the regression
coefficients from objects of class \code{'mvord'}.
}
