% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvord.R
\docType{data}
\name{data_cr_panel}
\alias{data_cr_panel}
\title{Simulated panel of credit ratings}
\format{A data frame with 11320 rows and 9 variables}
\usage{
data("data_cr_panel", package = "mvord")
}
\description{
A data set containing simulated credit ratings assigned by one rater and simulated perfomance measures for firms in different years.

\itemize{
  \item \code{rating} credit ratings
  \item \code{firm_id} firm index
  \item \code{year} year index
  \item \code{LR} liquidity ratio, relating the cash held by a company to the current liabilities
  \item \code{LEV} leverage ratio relating debt to earnings before interest and taxes
  \item \code{PR} profitability ratio of retained earnings to assets
  \item \code{RSIZE} log of relative size of the company in the market
  \item \code{BETA} a measure of systematic risk
  \item \code{BSEC} business sector of a firm (factor with 8 levels)
}
}
